Create .gitignore
When the start button is pressed, a randomly selected value is displayed in the top label.

Randomly selected keys must be displayed in the check boxes. No 2 check boxes may display the same key.

When the Mark Button is pressed, the application must check if the keys in the check boxes have the value displayed in the top label.

If there is a match and the user has checked the check box, a point is awarded. A point is also awarded should the key in the check box not match the value in the top label and the user has NOT checked the check box.

The user’s score must then be displayed in the bottom label out of 3.